[e2e] Rollback node affinity for BYOH label
This commit revert the node affinity logic introduced, in order to keep
the e2e-upgrade test low in complexity.

To ensure the presence of a running workload in a BYOH node, if any, we trust
the kube-scheduler placement algorithm, that balance the number of workloads
across the available nodes that match taints and selectors.

The number of workloads is set to 3, representing the total number of
nodes (one BYOH and one WM) + 1, assuring the eviction of at least 1 pod from
the BYOH node during the de-configuration process.

Follow-up to https://github.com/openshift/windows-machine-config-operator/commit/c85df6109de0af3485ae3583b98b27c207f06f19